Started implementing Generic CUDA
Adding lots of new code towards creating a generic CUDA implementation that automatically sets the correct parameters based on the detected GPU. Futhermore, it should also support the use of multiple GPUs. Since this is a big change to the current structure, it is implemented next to the existing CUDA implementations. Eventually, Maxwell and Kepler should be removed and HybridMaxwell should make use of the new Generic CUDA class.