Started to make PowerSensor work again for CUDA
The PowerRecord was previously declared in the CUDA class but is now moved to a seperate class. The callback mechanism does not work, so that no power readings are actually performed. This implementation is anyway a hack, since it requires a static PowerSensor instead of a class instance. A more thorough investigation is needed to make it work, preferably also with cleaner code.