CUDA power measurement using ArduinoPowerSensor is working again
To this end the ArduinoState had to be removed and merged with the State in PowerSensor. This is probably not the cleanest solution, since Arduino specific State members shouldn't be defined in the general PowerSensor header file. The same holds true for the Likwid specific State members. For now, this solution is at least much better than the previous implementation of PowerSensor. Todo is still the LikwidPowerSensor, which should derive from PowerSensor instead of being standalone.