Throw invalid argument exception when ArduinoPowerSensor cannot be initialized due to wrong POWER_DEVICE
For unknown reasons the split_string routine that is used to retrieve a list of power devices returned garbage when POWER_DEVICE was not set. It is therefore necessary to make sure that the ArduinoPowerSensor was indeed initialized correctly. If not, an DummyPowerSensor is used instead. This issue might also pop-up for the other proxies that use the ArduinoPowerSensor.