Fix degrid_visibilities throughput measurement
It is a bit of a hack, since the end timer does not wait for the whole queue to finish. This is due to the fact that the OpenCL runtime seems to postpone executing the remaining jobs for no good reason. Furthermore, the toggles have been moved from the header file to the source file, to prevent conflicts with proxies like HybridOpenCL.