Reimplemented powersensor
In this library, only a dummy power sensor is left, which does not
just measures runtime, not power consumption. All remaining powersensor
code is moved into a seperate repository and had to be linked (using
cmake) and specified (using POWER_SENSOR environmental variable)
explicitely. This gets rid of many dependencies that are not actually
needed to use IDG, while still supporting power measurements.