Updated synchronization mechanism for gridding routine
This routine still has a bug when using the wait() call for the last
event of a gridding iteration (job). This causes the device to
synchronize, reducing throughput. This call is used to report per-kernel
performance and can not be removed without additional code change. Since
profiling indicates that the compute and io are optimal without this
call, throughput is computed as a sum of all kernel execution times.