Reintroduced NR_CHANNELS as a compile-time constant
This makes the gridder kernel much simpler, as no template-like macro is
needed anymore and improves performance slightly. When a device is
initialized, the adder, splitter and scaler kernels are already compiled
(since they do not use the number of channels as argument). The gridder
and degridder kernels are compiled when they are first used, or when the
number of channels changes.