Restored C wrapper for util from idg-bin
This code needs additional cleanup. The typedefs are very ugly and some
of the routines exposed using this interface are not even used. Instead,
they are reimplemented in Python. This has to change so that either
Python-only initialization utilities are used, ore the ones provided in
idg-util (via the C interface and ctypes).