Initialize device memory for visibilities to zero
When visibilities are not gridded (they are discarded while creating the
Plan), they will naturally also not be computed in the degridder kernel.
Since the number of visibilities actually degridded can vary per
baseline, the degridding routine simply copies all visibilities for all
baselines. This includes some visibilities that are not initialized,
hence the entire buffer has to be initialized beforehand.