Initialize visibilities buffers prior to degridding
The degridder test compares the full visibilities buffers, including
visibilities that are not actually computed (because they were discarded
in while creating the Plan). The value for these visibilities should
not change during degridding. In order to be able to compare the two buffers
properly, they have to be initialized (to zero).