Update all submodules
These are changed for the new Data class, which provides an improved way
of initializing especially UVW data. This class is used in the updated
example programs (both cxx and python) to allow for example runs on
realistic, very large datasets. Some bugs that were encountered in the
proxies have been solved in the process.