Do a fftshift before and after the fft
fftshifts can be avoided if the images and grids are already in memory
such that the fft can be applied directly.

The previous state was very confusing because the fftshift was already
removed on one side of the fft, but still applied elsewhere
(in idg-api)

Now the shifts happen only in FFT.cpp, symmetrically on both sides of the fft

In the future the grids might be organized differently, requiring a change
in the indexing, such that the fftshifts can be eliminated. This should
then happen on both sides of the fft at once, to prevent confusion.