Run initialize, run_[de]gridding, finish_[de]gridding all on the same thread
CUDA proxies set a context in the initialize which is bound to the calling thread.
Subsequent CUDA calls need to happen from this thread.

Alternatively, the run_[de]gridding methods of the CUDA proxies need to set the context,
but that is not how it is currently implemented in idg-lib.