Fix out of memory error in cuMemAlloc called from GenericOptimized
out of memory error is probably caused by memory fragmentation, because
there appears to be sufficient memory.

The quick fix is to free the memory in each initialize, which prevents fragmentation,
but also disables the memory reuse in InstaceCUDA.

A bug in InstanceCUDA::free_device_memory() was also fixed. It freed the memory by
clearing a vector of raw pointers. This however does not free the memory.
The vector now contains std::unique_ptr elements, which ensures destruction when the
vector is cleared.