Reimplement Data class
A Data instance can now be created by specifying only a limit on the
number of stations and a limit on the baseline length. Grid size and
image size are moved to appropriate places. First, the user can compute
the image size that causes max uv-coverage for a grid of given size. Or
the other way around: compute the image size for max uv-coverage for a
grid of given size. Next, the image size is used to compute frequencies.
Finally, the frequencies and uvw coordinates in meters are used to
compute uvw in pixels (in Plan). This new Data implementation makes it
easier to test (realistic) uv-coverage.