Change compute/io overlap in degridding implementation
The asynchronous execution model has issues: pointers passed to the
routine could be invalid by the time that the asynchronous (memory copy) methods are invoked, causing
corruption, errors and/or crashes. Fall back to the same strategy as the
Generic proxy, where multiple host threads are used to overlap I/O and
computation.