Move aterm, avg_aterm_correction and spheroidal from KernelGridder into KernelGridderPost
By moving these post-processing operations out of the gridder kernel, it
becomes easier to optimize the gridding kernel. In this case an
additional pass from and to device memory is introduced, but the
overhead (in throughput) is neglible. The latencies for misaligned reads
(of aterm for instance) are easier to hide due to a higher occupancy.