Split checkCudaCall into checkCudaCall and assertCudaCall
assertCudaCall has the behavior of the previous checkCudaCall, it
checks whehter a CUDA function succeeded and throws and error in case
it did not. checkCudaCall prints, but does not pass on the error.

This serves as a (temporary) workaround for the issues encountered with
(un)registering host memory when using idg-api/wsclean.