Simplified GridderBufferImpl::flush_thread_worker
Creating plans and executing them is no longer overlapped. OpenMP causes
problems (such as bus errors) when nested parallelism is used in
combination with multiple parallel regions (one in GridderBufferImpl,
one in the GenericOptimized proxy and one in the CPU adder
kernel). In the current setting, having a simplier
flush_thread_worker method aids debugging and will not signicantly
affect performance of gridding as a whole.