Workaround: retry plan subgrid fft using smaller bulk size
It is not known exactly how much memory is needed for CUFFT. By default,
the jobsize takes this into account by only using say 80-90% of the
device memory. However, sometimes this is not sufficient. As a
workaround, the bulk size for the subgrid-fft is dynamically reduced
until a size is found where the fft is planned successfully.