Fix prediction with shifted phase centres
This commit solves three issues:

- In the bufferset, l was calculated from y and m from x. This is
fixed by swapping x and y.

- The definition of v and w within IDG are different than used
elsewhere, which results in that dl needs to be negated. I fixed
this by flipping the sign of 'l' when calling compute_n.

- For some unknown reason, the definition of l and m (or the
directions of the image) during stacking/unstacking is different
from those in the gridder/degridder. I worked around this by
calling compute_n in the stacking correction and gridder with
opposite signs of l and m. Together with the 2nd fix, this results
in calling compute_n with (-l, m) in the gridder and degridder,
and with (m, -l) in the stacking correction in BufferSet.

None of these changes have effect when dl=0 and dm=0, so this
does not affect standard wsclean imaging.