Optimize I/O by reusing a set of device buffers
This is an experimental commit, while the output is identical, something
weird is going on with the uvw coordinates: Cpying the uvw data during
initialization does not work, the uvw coordinates have to be copied just
before the kernel is executed. Furthermore, using the antenne_nr-th
d_uvw buffer does not work either. This might be related to the
synchronization issue between execution of the calibration kernel and
the copy of output. The problem areas are marked with FIXME in code.