Update do_calibrate_init and do_calibrate_update
Instead of preallocting device memory for aterm derivates, hessian and
gradient, these are now allocated in do_calibrate_update. The size of
these buffers is dependent on nr_terms, which is not known at the time
of initialization. Furthermore, copying hessian and gradient causes
runtime problems (hangs), these are now first copied in host buffers.