Add compute_residual method
This method computes the nr_terms+1-th term, which actually is the
residual visibility. This is needed such that nr_terms larger than
MAX_NR_TERMS can be iterated, while the residual visibility is already
needed for the first batch (not only the last batch untill nr_terms+1).
The current kernel has a lot of duplicate code and needs additional
cleanup.