Merge baseline_offset and time_offset in Metadata to time_index
The time_index specifies the first visibility (in time) associated with
the subgrid the Metadata describes. Previously, this had to be computed
from the baseline_offset and time_offset, with baseline_offset for the
first baseline possibly non-zero. In practice the Plan is always used
with some visibility buffer that should be processed from start to end,
hence, a time index in this buffer is simpler and suffices just as well.