Allocate d_sum_deriv and pass to KernelCalibrate
d_sum_deriv is used instead of d_scratch_sum, since it is used to store
the sums computed by combining the aterm for one station with the
derivatives for all terms. The new buffer d_sum_aterm is not used yet
but is intended to store the sums that are now computed in the
update_gradient method.