Workaround for GPU kernel compilation with old gnu compiler
Some C++ features are not supported prior to gcc/6. This is solved by
moving the Matrix2x2 and Visibility class from KernelTypes. to Types.h.
Furthemore, the index_X() methods are moved to a seperate header files.
This way, the CUDA kernels do not have to include the Types.h header
file that contains the C++ code.