Update to memory management in InstanceCUDA
Rename methods and functions to make a distinction between memory
allocation, memory registration (page-locking) and retrieving of
pre-allocated memory. All cu::HostMemory objects not wrapped in a
unique_ptr are contained in the h_misc_ vector, which uses a unique_ptr
to make sure that the memory is eventually released.