Integrate the Stopwatch class 'properly' into idg-lib
The Stopwatch uses Boost. To prevent Boost from becoming a dependency,
the boost components from the Stopwatch interface are moved into the
StopwatchImpl header, much like the BufferSet in idg-api. When Boost is
not available, the Stopwatch does not function. TODO: implement an
alternative, for instance based on OpenMP and a similar routine to
create a nice time string.