Reimplement Unified and UnifiedOptimized proxy
Call the gridding and degridding routines from the Generic Proxy instead
of duplicating a lot of code. Only the differences (tiling for both
proxies) and the FFT (for Unified) are maintained. Since tiling requires
a second grid anyway, allocate it internally and pass that grid onto the
routines. This saves a number of (large) memory copies.