Major cleanup of Generic and GenericOptimized proxies
Merge both (memory) initialization methods in a new CUDA::initialize
memory method. Only the things that are different between the two proxies
(host/device grids and subgrids) are implemented explicitely in the
respective proxy. Furthermore, the asynchronous code previously used
in the GenericOptimized proxy is replaced with code from the Generic
proxy. Finally, started to change some methods in InstanceCUDA to pass
the desired size in bytes for buffers rather than parameters. This
method is less error prone.