Reimplment CUDA Generic::do_gridding like GenericOptimized::do_gridding
The main objective is to get rid of the OpenMP loop that schedules work
by multiple threads in parallel. This mechnanism for unknown reason
causes the CUDA/CUFFT runtime to hang on specific machines.
The implementation in GenericOptimized avoids the use of OpenMP
alltogether. TODO: merge the Generic and GenericOptimized
implementations?