Enable rpath for libidg-api.so
This makes it easier to build software that uses idg (e.g. wsclean):
libidg-api.so now knows about the location of the CUDA toolkit used to
build idg-lib, such that the CUDA libraries do not need to be in the
LD_LIBRARY_PATH when building wsclean. The exception is libcuda.so,
which requires libcuda.so.1 which is only present when a NVIDIA driver
installed.