Add set_grid and get_grid methods to Proxy
These methods are intended for cases when the Proxy takes ownership over
the grid, for instance in the CUDA Generic proxy, where a copy of the
grid is kept in GPU memory. Only upon set_grid and get_grid, the host memory
is synchronized with the GPU memory. This approach has a number
consequences for the other CUDA proxies as well, hence a lot of changes
are made throughout the code.