Refactor ArrayTypes
No longer use a vector to store the dimensions of the array. Indexing
this vector frequently, for instance when indexing an array using the
() operator causes very significant overheads which need to be removed.
Without these vectors, much code must unfortunately be duplicated (like
it was with an earlier implementation of the Array types