Disable the option to build the OpenCL proxy and binaries
Two major changes from the 0.6 release break compatibility with the
OpenCL kernels (which have not been updated accordingly):
* Support for more than one aterm per subgrid
* Moved channel groups from BufferSet to Plan

Prior to releasing version 0.7, some changes to the (internal) API
prevent compilation of the OpenCL host-side code. Commit 82a2dd00 is the
last version where this code can be build, altough it can not be used
due to the aforementioned kernel incompatibilities.