Inital version of improved WSClean MPI wrapper
The improved script checks the status of the slurm queue for job completion
and fetches the exit status and/or signal from the slurm accounting.
The exit status of the script is the exit status of the slurm job that was
submitted. In case that the job was signalled by a Unix signal, the exit
status is 128+signal. This behaviour is similar to what the shell does.