CUDA W-Tiling: update wtile host addition in run_wtiles_to_grid
This code is now updated to add padded_tiles instead of w_padded_tiles
to the grid. In contrast to the w_padded_tiles, the padded_tiles are
scattered around d_tiles and have to be copied one by one to the host.
This code is tested for correctness, not yet for performance.  If this
proves to be slow, they could be first moved into a contigious buffer on
the GPU and than copied in bulk.