Fix a type issue for chunk_size
chunk_size was stored as int, but the chunk_size can be > signed 32 bit. On my local machine,
a 24 K x 24 K image caused it to overflow, whereas a 16 K x 16 K image seemed to work.

Seems like this was introduced in/around commit 59c367f13 , which
integrates the phase shift in CPU degridders.