Use index_grid() function in WTile kernels
This prevents int overflows in indexing in a standarized way.

Also fixes a sign flip in Stokes V when gridding with WTiling.
Integration tests currently do not test Stokes V.
Manually running the tGridDegrid test with
"V" in the STOKESLIST in test_gridding.py confirms
that this was broken and fixed now.