CUDA W-Tiling: add test-cuda-wtiling-forward.x
This tests the CUDA W-Tiling code on the InstanceCUDA level and therfore
serves as a unit test, whereas test-hybrid-cuda-unified.x tests W-Tiling
on the proxy level. The latter test fails for unknown reason, while this
new test demonstrates that the CUDA kernels are at least correct when
used as units.