Add InstanceCUDA::set_powersensor
This is much cleaner than passing both device_nr and device_id in its
constructor. InstanceCUDA relies on CUDA to call this method, otherwise
the power sensor will be uninitialised. This will be fixed in a future
commit when the raw pointer to the power sensor is replaced by a smart
pointer.