Let all CUDA routines manage their own device memory
This means that m_buffers had to be removed from CUDA, as well as the
compute_jobsize and initialize methods. These changes has led to some
duplicated code, but it is now much more clear what method uses what
memory. A new (much simpler) initialize_jobs method computes the jobsize
and initializes a vector of JobData as before.