Fix tile batch size and tile fft
The jobs had a static batch size, which is incompatible with the fft
requiring memory based on the w padded tile size. Without jobs, we can
simply increment the tile offset with the actual number of tiles
processed in a batch. As an added bonus, only a single call to
plan_tile_fft is needed.