Improve Dockerfile
Improved the Dockerfile for building a base image for Rapthor. It now more closely follows the approach taken by Prefactor3-CWL: have two Dockerfiles. One for building the base image, and one for the complete
image. The base image can be used in CI for testing against a checked out repository. The complete image can be used to create a distributable container image.