WIP: Let rapthor run with Singularity
Let the rapthor pipeline run using Singularity. This requires changes to
the CWL steps, but also to the code that generates the pipeline_parset.cwl
and pipeline_inputs.yml files. The debug directory contains a run.sh script
that can be used to first figure out what changes need to be made, both to
the CWL script files and to the code that generates the pipeline_parset.cwl
and pipeline_inputs.yml files. This is very much work in progress, and the
pipeline currently does _not_ run successfully yet.