Fix error in imaging pipeline
It seems we need the MultipleInputFeatureRequirement in the imaging
pipeline. Not sure why this problem didn't show up earlier. It may
have to do with the fact that the imaging pipeline is using array
of array of File/Directory. At least, that's what the description of
MultipleInputFeatureRequirement is hinting at. However, neither David
nor me really understand what this requirement does exactly ;)