Fix WSClean exception writing temporary file
In many of the CWL steps that invoke WSClean, the argument to `-temp-dir`
was given as user-specified input (a path pointing to the scratch
directory used by Rapthor). Problem is that this directory is not
mounted inside a Docker container, hence cannot be written to.
This fix uses the `$(runtime.tmpdir)` CWL variable to determine the
location of the temporary directory at runtime. Its value can be managed
by the CWLTool/Toil command-line option `--tmpdir-prefix`. Rapthor sets
this option based on the value of `dir_local` in the parset file.