Fix memory usage of subtract step
The script tries to fit things in memory by running `du`, but because the dest is a symlink, it returns 0. This causes the script to use only one chunk, which requires an enormous amount of memory. The fix is to add -L to du, so that it dereferences the symlink.