Modify concat script
Modified the `concat_linc_files` script. This was necessary, because the input files, when fetched by the `fetch_data.cwl` CWL step, will not be put in a single directory.
The new `concat_ms.py` script now contains most of the logic that used to be in `concat_linc_files`. This script can handle multiple inputs MS paths. The API and behavior of `concat_linc_files` haven't changed.