Handle optional arguments properly
Properly handle optional arguments that can be passed as input to this CWL step. More specifically, make sure that absent optional inputs are removed from the input JSON, so that they do not end up in the generated parset file, but at the same time ensure that they are updated and _do_ end up in the parset file if they are present.