Revert "Use setuptools_scm"
Python 3.6 does not proper deal with `pyproject.toml`. As a result, it ignores the Python >= 3.7 requirement, and `pip install .` fails to determine the package name, causing it to install a package named `UNKNOWN`.
Since this change is not essential for our first release, I reverted it. This gives me more time to figure out how to make this also work for Python 3.6, in the sense that it should simply fail to install at all.

This reverts commit 7ded4da8966064c09ddbc3f198a319330047096a.