Let CI use Docker/Dockerfile
Currently, CI uses a `ci/ubuntu_20_04-base` to build the base image.
The `Docker` directory contains a `Dockerfile` and some helper scripts to build a docker image.
It's cumbersome and error prone to maintain two different files for building a docker image, so let CI use the one in the `Docker` directory.