Always bypass file store
The use of the Toil file store gives rise to issues when used in combination with Singularity. In that case the chain of symbolic links cannot be completely dereferenced, because some (intermediate) paths are not mounted inside the container. Bypassing the file store solves that problem. There is no real benefit in using a file store, because Rapthor does not use an object store like S3. Therefore, Rapthor will now _always_ bypass the Toil file store.

Fixes RAP-453