Ensure MS exists for test_parset.py
The MS `resources/test.ms` is not guaranteed to exist when running `test_parset.py`.
It is created by `test_field.py` if it doesn't exist yet. So, if `test_parset.py` is run before `test_field.py`, it will fail.
Therefore, we now create a `dummy.ms` which _is_ guaranteed to exist.