Refactor `parset.py`.
Before the refactoring, all the logic to read a Rapthor parset file and to handle the different options resided in a single `parset_read` function. As part of the refactoring, a new `Parset` class was created that now handles the reading of a parset file, the validation of its contents, and the conversion to a `dict`. The remaining part of the logic in the original `parset_read` function dealt with preparing the runtime environment, e.g., by creating a working directory, checking for the presence of input MS files, etc. That part has been retained, though you can debate if `parset_read` is the proper name for such a function.