Fix missing close quote
The proper fix is in the function `add_coloring_to_emit_ansi` in `_logging.py`.
For reasons that were sensible at that time (I guess), any trailing single quotes were stripped from the exception message.
This was done to avoid error strings like `KeyError: 'missing_key'`, but I think it's better to not modify this string.