Fix issue with plot_rapthor_timing script
After Rapthor learned the new `initial_image` command, the script to plot timings broke. Reason is that the `initial_image` command is executed only once, and doesn't have a trailing `_<cycle-nr>` in the directory name.
The solution for now is to add a trailing `_1` to those operations (currently only `initial_image`) that do not have one.