Alternative implementation for RAP-973
This implementation is different in the sense that it puts temporary data in a unique sub-directory when using Toil in single-machine mode. This will ensure that, if two or more instances of Rapthor are running concurrently, each Rapthor instance will only remove the temporary files/directories it created itself, and will not clobber other data created in the scratch directories.