Revert to the version of March 11
Reverted to the version of March 11, which did not yet use a unique sub-directory for temporary output when using Toil in single-machine mode.
Rationale: this solution, though it suffers from a race-condition, is cleaner; and we do not really need to be able to run Rapthor concurrently with the same settings for `local_scratch_dir` and `global_scratch_dir`.
TODO: add a warning about this in the documentation.