Mark all CWL workflow tests as slow
The tests in `test_cwl.py` test a large variety of generated CWL workflows. Testing every combination is very slow. The whole test module is therefore marked as 'slow'. By default, slow tests will not be run, because the option `-m "not slow"` was added to the default test environment invoked by `tox`.