Run test from tests directory
Make sure that all the tests in `test_field` are run from the `tests` directory
TODO: Find a better way (e.g., using fixtures) to make sure resources can be found. Currently, we need to be in the `tests` directory, because the test(s) use a relative path to the resources.